+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Session and/or refresh tokens can now be disabled with a client logout.
 - The client now supports session auto-refresh using refresh tokens. This is enabled by default.
 - New socket RPC and MatchSend methods using ArraySegment to allow developers to manage memory re-use.
+- Add IAP validation APIs for purchase receipts with Apple App Store, Google Play Store, and Huawei AppGallery.
 
 ## [2.8.0] - 2020-02-19
 ### Changed